+++ app/Http/Controllers/FournisseurController.php
@@ -1,16 +1,21 @@
 <?php
 
+// adapter le controller pour prendre en compte l'entreprise
+
 namespace App\Http\Controllers;
 
 use App\Models\Fournisseur;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class FournisseurController extends Controller
 {
 
+    // TODO : Adapter la fonction index pour prendre en compte l'entreprise
     public function index()
     {
-        $fournisseurs = Fournisseur::withTrashed()->get();
+        $en_id = Auth::user()->entreprise->en_id;
+        $fournisseurs = Fournisseur::where('en_id', $en_id)->get();
         return view('pages.fournisseur.index', compact('fournisseurs'));
     }
 
@@ -19,42 +24,52 @@ class FournisseurController extends Controller
         return view('pages.fournisseur.create');
     }
 
+    // TODO : Adapter la fonction store pour prendre en compte l'entreprise
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
+        $data = $request->validate([
             'f_nom' => ['required', 'max:100'],
-            'f_tel' => ['required', 'max:25'],
-            'f_adr' => ['required', 'min:10'],
+            'f_tel' => ['max:25'],
+            'f_adr' => '',
         ]);
-        Fournisseur::create($validatedData);
+
+        $en_id = Auth::user()->entreprise->en_id;
+        $data['en_id'] = $en_id;
+
+        Fournisseur::create($data);
         return redirect()->route('fournisseur.index')->with('success', 'Fournisseur créé avec succès');
     }
 
-    public function show(Fournisseur $fournisseur)
-    {
-        return view('pages.fournisseur.show', compact('fournisseur'));
-    }
 
+    // TODO : Adapter la fonction edit pour prendre en compte l'entreprise
     public function edit(Fournisseur $fournisseur)
     {
+        // empecher a d'autres de modifier les données d'une autre entreprise
+        if ($fournisseur->en_id != Auth::user()->entreprise->en_id) {
+            return abort(403, 'Vous n\'êtes pas autorisé à modifier ce fournisseur');
+        }
         return view('pages.fournisseur.edit', compact('fournisseur'));
     }
 
     public function update(Request $request, Fournisseur $fournisseur)
     {
-        $validatedData = $request->validate([
+
+        $data = $request->validate([
             'f_nom' => ['required', 'max:100'],
-            'f_tel' => ['required', 'max:25'],
-            'f_adr' => ['required'],
+            'f_tel' => ['max:25'],
+            'f_adr' => '',
         ]);
-
-        $fournisseur->update($validatedData);
-
+        $fournisseur->update($data);
         return redirect()->route('fournisseur.index')->with('success', 'Fournisseur mis à jour avec succès');
     }
 
+    // TODO : Adapter la fonction destroy pour prendre en compte l'entreprise
     public function destroy(Fournisseur $fournisseur)
     {
+        // empecher a d'autres de supprimer les données d'une autre entreprise
+        if ($fournisseur->en_id != Auth::user()->entreprise->en_id) {
+            return abort(403, 'Vous n\'êtes pas autorisé à supprimer ce fournisseur');
+        }
         $fournisseur->delete();
         return redirect()->route('fournisseur.index')->with('success', 'Fournisseur supprimé avec succès');
     }
